Fixing logic and starting on tests

+++ backend/src/main/kotlin/ntnu/idatt2105/statistics/dto/StatsDto.kt
@@ -1,7 +1,8 @@
 package ntnu.idatt2105.statistics.dto
 
-data class StatsDto(val nrOfReservation: Int,
-                    val hoursOfReservation: Double,
-                    val daysWithReservation: Int,
-                    val userReservationCount: Int,
+data class StatsDto(
+    val nrOfReservation: Int,
+    val hoursOfReservation: Long,
+    val daysWithReservation: Int,
+    val userReservationCount: Int,
                     )
